Move time block files into CreateSchedule directory and refactor TimeBlockForm.css so its not too big
