--- frontend/src/components/stylesheets/TimeBlockForm.css
@@ -1,273 +1,0 @@
-.time-block-form-btn {
-    display: flex;
-    gap: var(--space-lg);
-    margin-top: var(--space-xl);
-}
-
-.btn:hover {
-    box-shadow: var(--shadow-xl);
-    border-color: var(--border-soft);
-}
-
-.time-block-form-card {
-    border: 1px solid var(--border-soft);
-    width: min(900px, 95vw);
-    border-radius: var(--radius-lg);
-    box-shadow: var(--shadow-lg);
-    padding: var(--space-xl);
-    margin: var(--space-md);
-    text-align: center;
-    background: var(--bg-soft);
-    color: var(--text-main);
-}
-
-.time-block-section {
-    display: flex;
-    flex-direction: column;
-    gap: var(--space-xs);
-    margin: var(--space-2xl) 0;
-    padding: var(--space-xl);
-    border: 1px solid var(--border-soft);
-    border-radius: var(--radius-lg);
-    box-shadow: var(--shadow-lg);
-    background: var(--bg-main);
-}
-
-.page-center {
-    min-height: 100vh;
-    width: 100vw;
-    display: flex;
-    justify-content: center;
-    align-items: flex-start;
-    padding: var(--space-3xl) var(--space-lg);
-    overflow-y: auto;
-    background: var(--bg-main);
-    color: var(--text-main);
-}
-
-.delete-btn {
-    display: flex;
-    gap: var(--space-lg);
-    margin-top: var(--space-xl);
-    background-color: var(--danger);
-    color: var(--text-white);
-    border-radius: var(--radius-lg);
-    width: 125px;
-}
-
-.btn.cancel-btn {
-    background-color: var(--blue);
-    color: var(--text-white);
-    border: 1px solid var(--blue);
-}
-
-.btn.cancel-btn:hover {
-    background-color: var(--blue);
-    color: var(--text-white);
-}
-
-.btn.edit-btn {
-    background-color: var(--edit);
-    color: var(--text-white);
-    border: 1px solid var(--edit);
-}
-
-.btn.edit-btn:hover {
-    background-color: var(--edit-soft);
-    color: var(--text-main);
-}
-
-.error-text-date {
-    color: var(--danger);
-    font-size: var(--font-size-sm);
-    margin-top: 2px;
-    margin-bottom: 2px;
-}
-
-.error-text {
-    color: var(--danger);
-    font-size: var(--font-size-sm);
-    margin-top: 2px;
-    margin-bottom: 2px;
-    text-align: left;
-}
-
-.tab-bar {
-    display: flex;
-    border-bottom: 2px solid var(--border-light);
-    margin-bottom: var(--space-2xl);
-    gap: 0;
-}
-
-.tab-btn {
-    padding: 0.6rem 1.25rem;
-    background: none;
-    border: none;
-    border-bottom: 2px solid transparent;
-    margin-bottom: -2px;
-    cursor: pointer;
-    font-size: 0.95rem;
-    color: var(--text-muted);
-    transition:
-        color 0.15s,
-        border-color 0.15s;
-}
-
-.tab-btn:hover {
-    color: var(--text-main);
-}
-
-.tab-btn--active {
-    color: var(--blue);
-    border-bottom-color: var(--blue);
-    font-weight: 600;
-}
-
-.tab-panel {
-    padding: var(--space-xs) 0;
-}
-
-.global-options {
-    display: flex;
-    gap: var(--space-2xl);
-    margin: 1rem 0;
-    padding-left: var(--space-2xl);
-}
-
-.checkbox-label {
-    display: flex;
-    align-items: center;
-    gap: 0.4rem;
-    cursor: pointer;
-    font-size: 0.95rem;
-    color: var(--text-main);
-    
-}
-
-.checkbox-label--disabled {
-    opacity: 0.4;
-    cursor: not-allowed;
-}
-
-.checkbox-label-input:checked {
-    background-color: var(--blue);
-    border-color: var(--blue);
-    background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3E%3Cpath fill='none' stroke='white' stroke-width='2' d='M3 8l4 4 6-6'/%3E%3C/svg%3E");
-    background-repeat: no-repeat;
-    background-position: center;
-    background-size: var(--font-size-xs);
-}
-
-.checkbox-label-input {
-    flex-shrink: 0;
-    margin: 0 !important;
-    width: var(--checkbox-size);
-    height: var(--checkbox-size);
-    position: static !important;
-    appearance: none;
-    -webkit-appearance: none;
-    background-color: var(--bg-main);
-    border: 2px solid var(--border-light);
-    border-radius: var(--radius-sm);
-    cursor: pointer;
-    transition: background-color var(--transition-speed), border-color var(--transition-speed);
-}
-
-.spinner {
-    display: inline-block;
-    width: 14px;
-    height: 14px;
-    border: 2px solid rgba(255, 255, 255, 0.4);
-    border-top-color: white;
-    border-radius: 50%;
-    animation: spin 0.7s linear infinite;
-    margin-right: var(--space-sm);
-    vertical-align: middle;
-}
-
-@keyframes spin {
-    to {
-        transform: rotate(360deg);
-    }
-}
-
-.range-box {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-    gap: var(--space-md);
-    justify-content: center;
-    align-items: end;
-    margin-bottom: var(--space-sm);
-}
-
-.range-box label {
-    display: flex;
-    flex-direction: column;
-    color: var(--text-main);
-    min-width: 0;
-    flex-wrap: wrap;
-    align-content: space-around;
-    justify-content: space-evenly;
-    align-items: flex-start;
-    gap: var(--space-xs);
-}
-
-.range-box input {
-    width: 100%;
-    min-width: 0;
-    max-width: 260px;
-}
-
-.time-block-form-card input,
-.time-block-form-card select,
-.time-block-form-card textarea {
-    padding: var(--space-sm) 14px;
-    border-radius: var(--radius-lg);
-    border: 1px solid var(--border-input);
-    background-color: var(--bg-main);
-    font-size: var(--font-size-sm);
-    color: var(--text-main);
-    box-shadow: var(--shadow-sm);
-    transition: all 0.2s ease;
-}
-
-.time-block-form-card input::placeholder,
-.time-block-form-card textarea::placeholder {
-    color: var(--text-muted);
-    opacity: 1;
-}
-
-.time-block-form-card input:focus,
-.time-block-form-card select:focus,
-.time-block-form-card textarea:focus {
-    outline: none;
-    border-color: var(--blue);
-    box-shadow: var(--shadow-focus);
-}
-
-.time-block-form-card input[type="checkbox"] {
-    border-radius: 5px;
-    width: 20px;
-    height: 20px;
-    padding: 0;
-    transition: all 0.1s ease;
-}
-
-.time-row {
-    display: flex;
-    align-items: center;
-    gap: var(--space-lg); 
-    width: 100%;
-}
-
-.time-row label,
-.date-row label {
-    min-width: var(--space-3xl); 
-}
-
-.time-row input[type="time"] {
-    flex: 1;        
-    width: 100%;  
-    padding: var(--space-md) var(--space-lg);
-    border-radius: var(--radius-lg);
-}
