+++ frontend/src/functionality/CreateSchedule/CreateSchedule.jsx
@@ -1,8 +1,9 @@
 import { useState } from 'react';
-import TimeBlockForm from '../../components/TimeBlockForm.jsx';
+import TimeBlockForm from './TimeBlockForm.jsx';
 import GeneratorForm from './GeneratorForm.jsx';
 import { useCreateTimeBlock } from './utils/Hooks/useCreateTimeBlock.js';
 import { useGenerateSchedule } from './utils/Hooks/useGenerateSchedule.js';
+import './stylesheets/CreateSchedule.css';
 
 const TABS = [
     { id: 'time-block', label: 'Task' },
