+++ frontend/src/functionality/CreateSchedule/FormActions.jsx
@@ -1,3 +1,5 @@
+import './stylesheets/FormActions.css';
+
 /**
  * Handles all action buttons for the TimeBlockForm.
  * - Shows "Add Another Event" only when creating new blocks
