+++ frontend/src/functionality/CreateSchedule/GeneratorForm.jsx
@@ -1,11 +1,12 @@
-import '../../components/stylesheets/TimeBlockForm.css';
 import { BLOCK_TYPES } from '../../constants/blockTypes.js';
 import GeneratorFormItem from './GeneratorFormItem.jsx';
 import DateRangeInput from './DateRangeInput.jsx';
 import TimeWindowInput from './TimeWindowInput.jsx';
 import GlobalOptions from './GlobalOptions.jsx';
-import FormActions from '../../components/FormActions.jsx';
+import FormActions from './FormActions.jsx';
 import useGeneratorForm from '../../utils/Hooks/useGeneratorForm.js';
+import './stylesheets/GeneratorForm.css';
+
 
 /**
  * GeneratorForm renders schedule input fields and manages unscheduled block state.
