+++ frontend/src/functionality/CreateSchedule/stylesheets/GeneratorForm.css
@@ -1,0 +1,71 @@
+.global-options {
+    display: flex;
+    gap: var(--space-2xl);
+    margin: 1rem 0;
+    padding-left: var(--space-2xl);
+}
+
+.checkbox-label {
+    display: flex;
+    align-items: center;
+    gap: 0.4rem;
+    cursor: pointer;
+    font-size: 0.95rem;
+    color: var(--text-main);
+}
+
+.checkbox-label--disabled {
+    opacity: 0.4;
+    cursor: not-allowed;
+}
+
+.checkbox-label-input {
+    flex-shrink: 0;
+    margin: 0 !important;
+    width: var(--checkbox-size);
+    height: var(--checkbox-size);
+    position: static !important;
+    appearance: none;
+    -webkit-appearance: none;
+    background-color: var(--bg-main);
+    border: 2px solid var(--border-light);
+    border-radius: var(--radius-sm);
+    cursor: pointer;
+    transition: background-color var(--transition-speed), border-color var(--transition-speed);
+}
+
+.checkbox-label-input:checked {
+    background-color: var(--blue);
+    border-color: var(--blue);
+    background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3E%3Cpath fill='none' stroke='white' stroke-width='2' d='M3 8l4 4 6-6'/%3E%3C/svg%3E");
+    background-repeat: no-repeat;
+    background-position: center;
+    background-size: var(--font-size-xs);
+}
+
+.range-box {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+    gap: var(--space-md);
+    justify-content: center;
+    align-items: end;
+    margin-bottom: var(--space-sm);
+}
+
+.range-box label {
+    display: flex;
+    flex-direction: column;
+    color: var(--text-main);
+    min-width: 0;
+    flex-wrap: wrap;
+    align-content: space-around;
+    justify-content: space-evenly;
+    align-items: flex-start;
+    gap: var(--space-xs);
+}
+
+.range-box input {
+    width: 100%;
+    min-width: 0;
+    max-width: 260px;
+}
