+++ frontend/src/functionality/CreateSchedule/stylesheets/SuccessfulTimeBlock.css
