+++ frontend/src/functionality/CreateSchedule/stylesheets/TimeBlockItem.css
@@ -1,0 +1,21 @@
+.time-block-section {
+    display: flex;
+    flex-direction: column;
+    gap: var(--space-xs);
+    margin: var(--space-2xl) 0;
+    padding: var(--space-xl);
+    border: 1px solid var(--border-soft);
+    border-radius: var(--radius-lg);
+    box-shadow: var(--shadow-lg);
+    background: var(--bg-main);
+}
+
+.delete-btn {
+    display: flex;
+    gap: var(--space-lg);
+    margin-top: var(--space-xl);
+    background-color: var(--danger);
+    color: var(--text-white);
+    border-radius: var(--radius-lg);
+    width: 125px;
+}
